chore(main): clean up routes (#879)
* chore(main): cleanup routes

* chore(main): missing routes
