+++ client/js/pages/file-manager/file.actions.js
@@ -3,10 +3,31 @@ export const doesFileExist = () => {
 
 }
 
-export const createFile = () => {
+export const overWriteFile = () => {
 
+}
+
+export const upRevisionFile = () => {
+
+}
+
+export const createFile = () => {
+    //does file exist
+
+    //if exists - overwrite or up-rev
+
+    // overwrite file
+
+    // up-rev file
+
+    //else upload file...
 }
 
 export const moveFile = () => {
 
 }
+
+export const deleteFile = () => {
+
+}
+
