+++ server/views/dashboard.ejs
@@ -122,7 +122,7 @@
         <% if(dashboardSettings?.historical.environmentalHistory){ %>
             <%- include(itemPath, {
             tile:{
-            w:12,h:6,mw:3,mh:6,maxh:10,id:"enviroData",bodyId:"environmentalHistory",title:"Environmental History",i:"fa-tree", c:"pt-2 text-center"
+            w:12,h:6,mw:3,mh:6,maxh:10,id:"enviroData",bodyId:"environmentalHistory",title:"Environmental HistoryRoutes",i:"fa-tree", c:"pt-2 text-center"
             },
             titleFile:'./environment-alert.title.ejs'
             })
