+++ server/views/history.ejs
@@ -1011,7 +1011,7 @@
   <div class="row mt-5">
     <div class="col-md-6 m-auto">
       <div class="card card-body text-center">
-        <h1>No History yet</h1>
+        <h1>No HistoryRoutes yet</h1>
         <p>
           Come back here when you've completed some prints.
         </p>
