+++ server/views/layout.ejs
@@ -146,8 +146,8 @@
                 </a>
             </li>
             <li class="nav-item">
-                <a title="Print History" class="nav-link text-center" href="/history" id="historyBtn">
-                    <i class="fas fa-history"></i> History
+                <a title="Print HistoryRoutes" class="nav-link text-center" href="/history" id="historyBtn">
+                    <i class="fas fa-history"></i> HistoryRoutes
                 </a>
             </li>
             <li class="nav-item">
