+++ server/views/partials/PrinterManager/healthChecks.modal.ejs
@@ -47,7 +47,7 @@
                     <p>Checks to see if your printer webcam settings are setup correctly and correspond to what data
                         OctoFarm has.</p>
                     <hr>
-                    <p><b>History Settings</b></p>
+                    <p><b>HistoryRoutes Settings</b></p>
                     <p>Checks to see if OctoPrints settings are correctly setup for OctoFarms history collection
                         settings.</p>
                     <hr>
@@ -77,7 +77,7 @@
                 <th scope="col" class="sticky-table table-dark" style="">Printer Profile</th>
                 <!--                                Webcam Settings-->
                 <th scope="col" class="sticky-table table-dark" style="">Webcam Settings</th>
-                <th scope="col" class="sticky-table table-dark" style="">History Collection Settings</th>
+                <th scope="col" class="sticky-table table-dark" style="">HistoryRoutes Collection Settings</th>
             </tr>
             </thead>
             <tbody id="healthChecksTable" style="height:100%; overflow-y:auto;">
