+++ server/views/partials/System/systemClient.ejs
@@ -392,7 +392,7 @@
                                            id="environmentalHistory" required>
                                     <label class="custom-control-label"
                                            for="environmentalHistory">Environmental
-                                        History</label><br>
+                                        HistoryRoutes</label><br>
                                     <div class="valid-feedback">Shown</div>
                                     <div class="invalid-feedback">Hidden</div>
                                 </div>
