+++ server/views/partials/System/systemDatabase.ejs
@@ -32,7 +32,7 @@
     </button>
     <button id="exportHistory" type="button"
             class="btn btn-danger btn-lg bg-colour-4 mb-2"><i class="fas fa-history"></i>
-        <br> Export History
+        <br> Export HistoryRoutes
     </button>
     <button id="exportPrinters" type="button"
             class="btn btn-warning btn-lg bg-colour-1 mb-2"><i
@@ -74,7 +74,7 @@
     </button>
     <button id="nukeHistory" type="button"
             class="btn btn-danger btn-lg bg-colour-4 mb-2"><i class="fas fa-history"></i>
-        <br> Nuke History
+        <br> Nuke HistoryRoutes
     </button>
     <button id="nukePrinters" type="button"
             class="btn btn-warning btn-lg bg-colour-1 mb-2"><i
