+++ server/views/partials/System/systemServer.ejs
@@ -103,7 +103,7 @@
             <a class="nav-link"
                id="server-history-list" data-toggle="list" href="#server-history"
                role="tab" aria-controls="listview"><i class="fas fa-history"></i>
-                History</a>
+                HistoryRoutes</a>
         </li>
         <li class="nav-item">
             <a class="nav-link"
@@ -506,7 +506,7 @@
                  aria-labelledby="server-history-list">
                 <div class="alert alert-secondary" role="alert">
                     OctoFarm has the ability to save a snapshot/thumbnail into
-                    History. These options are turned on by default as it will check
+                    HistoryRoutes. These options are turned on by default as it will check
                     if the thumbnail/camera url exist before actually saving.
                 </div>
                 <div class="alert alert-warning text-dark text-dark" role="alert">
@@ -682,7 +682,7 @@
                                 data-target="#collapseExample2"
                                 aria-expanded="false"
                                 aria-controls="collapseExample2">
-                            History Information
+                            HistoryRoutes Information
                         </button>
                         <button class="btn btn-primary" type="button"
                                 data-toggle="collapse"
@@ -751,7 +751,7 @@
                     <div class="collapse" id="collapseExample2"
                          data-parent="#influxParent">
                         <div class="alert alert-dark" role="alert">
-                            History information will be sent when a print is logged
+                            HistoryRoutes information will be sent when a print is logged
                             as completed or failed (cancelled / error).
                         </div>
                         <div class="alert alert-warning text-dark text-dark" role="alert">
